+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.10.0] ???
+### Added
+- Added `non_retrogradable_measures` function to `utils.py` for finding all palindromic measures in a given filepath and part number. 
 
 ## [v0.9.1] April 16, 2021
 ### Added